fix: Room destroyed during join
If a User was in the process of joining while the Admin killed the room, the User would get stuck in a state where no room existed but no messaging was displayed.

Now the Client and Server will inform the User if the room was taken down during the Join Room phase.