fix: Selected card not toggleable
When I moved all the game logic to the Server, there was a miss with the toggleable Cards.
So if two cards are selected for answers, and a User tried to deselect one of the two, the `ndx` wasn't set on the selected cards (since it was being set on the Client only).

The `ndx` of cards will now be handled on the Server.