feat: Show User as disconnected
Sometimes a User may get temporarily disconnected from a game. Previosly I was just giving them 1 second to reconnect before considering them as leaving the game. Now it'll wait 5 seconds which should be plenty of time for a reload, and that User will be displayed to everyone as disconnected before eventually being bounced from the room.