fix: JSON stringify error
Was getting a "Converting circular structure to JSON" error during Server restart. Determined it only happened after I had Users in a game, and I updated Client code which then triggered a sync.
The cause was that the new `disconnectCheck` I added, was bleeding over in some instances, and you can't serialize a Timeout of course. Just had to refactor some things so that logic lives in action's context and not the data.