fix: Card remains in selected state after Czar left
If a User had a card selected (but not submitted), and the Czar left the game, then a second User joins and is made the Czar, then the User that previously selected a card but didn't submit their answer would see that previously selected card still showing up as selected (but not marked as submitted by the game state).

Now cards will be reset to their default state when cards are dealt.