fix: Error on Home page after disconnect
There was a race condition with the Modal after a disconnect, but only on the Home page. Depending on the nesting structure of the DOM, sometimes there'll be something for the Modal to delete when `forceClose` is called, and sometimes not. I already had a gate to not try and delete a Modal if it wasn't in the Portal, but there was extra Svelte cleanup code that was executing after I'd already cleared the DOM, and it was failing because I deleted what it was trying to delete.
Now I wait for Svelte to do it's thing, then I try to close any open Modals.