fix: Sound plays when page not focused
If the User switched to another page while waiting, and they get awarded points, the sound would play but after re-focusing the page the animation would play.
Just added some checks to verify the page is focused and that the sound hasn't already played (in case they switch back and forth without closing the dialog first).